[grafana] Tighten bridge docstrings, tests, and module layout
Merge results.py into server.py, fold the four bridge test files into one
e2e test_bridge.py, and reword docstrings and comments to state behavior
plainly. No behavior change.
