+++ infra/grafana/src/config.py
@@ -1,34 +1,28 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Bridge configuration, resolved once at startup from the container environment.
+"""Bridge configuration: the clusters served and the runtime settings.
 
-One bridge serves every cluster in :data:`CLUSTERS`; Grafana provisions one
-datasource per entry and addresses it by name in the URL path. Both finelog VMs
-live in the same project and zone and are reached over Direct VPC egress on their
-internal IPs, so a single Cloud Run service covers prod and dev.
-
-Cluster targets are defined here and changed under review, so a datasource points
-at the cluster it names.
+One bridge serves every cluster in CLUSTERS. Grafana provisions one datasource
+per entry and addresses it by name in the URL path.
 """
 
 import dataclasses
 import os
 
-# finelog listens here on both VMs (lib/finelog/config/{marin,marin-dev}.yaml).
+# Port finelog listens on, set in lib/finelog/config/{marin,marin-dev}.yaml.
 FINELOG_PORT = 10001
 
-# The bridge's loopback port, named literally by the provisioned datasource URLs
-# (provisioning/datasources/finelog.yaml). A constant so the two cannot drift.
+# Loopback port the bridge listens on; the provisioned datasource URLs
+# (provisioning/datasources/finelog.yaml) use it.
 BRIDGE_PORT = 8081
 
 
 @dataclasses.dataclass(frozen=True)
 class ClusterTarget:
-    """One finelog deployment the bridge can query.
+    """A finelog deployment the bridge can query.
 
-    ``instance_filter`` is a GCE list filter selecting the VM whose internal IP
-    the bridge connects to.
+    instance_filter is a GCE list filter selecting the VM to connect to.
     """
 
     name: str
@@ -57,19 +51,15 @@ CLUSTERS: tuple[ClusterTarget, ...] = (
 class BridgeConfig:
     """Resolved bridge settings."""
 
-    # Rows one query may return. finelog caps a response at 64 MiB and enforces its
-    # own query deadline; this is a lower ceiling so a mis-written panel returns an
-    # error rather than a result Grafana cannot render.
+    # Maximum rows a query may return before the bridge rejects it with a 400.
     max_rows: int
-    # Result cache TTL. Grafana's own query caching is Enterprise-only, and a
-    # shared dashboard auto-refreshing across viewers would otherwise multiply
-    # straight through to the finelog hub.
+    # Result cache TTL, seconds.
     cache_ttl: float
     query_timeout_ms: int
 
     @staticmethod
     def from_environment() -> "BridgeConfig":
-        """Read settings from the container env, falling back to prod-safe defaults."""
+        """Read settings from the container environment."""
         return BridgeConfig(
             max_rows=int(os.environ.get("GRAFANA_BRIDGE_MAX_ROWS", "200000")),
             cache_ttl=float(os.environ.get("GRAFANA_BRIDGE_CACHE_TTL", "20")),
