+++ infra/grafana/tests/test_bridge.py
@@ -1,13 +1,15 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Behavioural tests for the bridge's HTTP surface, against a fake finelog."""
+"""Tests for the grafana bridge: its HTTP surface over a fake finelog, and the
+cache's coalescing and eviction contract."""
 
+import threading
 from datetime import UTC, datetime
 
 import pyarrow as pa
+from cache import TtlCache
 from config import BridgeConfig, ClusterTarget
-from conftest import finelog_result
 from finelog.errors import QueryResultTooLargeError
 from server import create_app
 from starlette.testclient import TestClient
@@ -15,10 +17,14 @@ from starlette.testclient import TestClient
 # 2026-07-17T03:00:00Z and +1h, as Grafana sends them.
 FROM_MS = 1_784_257_200_000
 TO_MS = FROM_MS + 3_600_000
-
 MARIN = ClusterTarget(name="marin", project="p", zone="z", instance_filter="name = finelog-marin")
 
-# A canned result with a timestamp column, so a query returns plottable rows.
+
+def finelog_result(**columns: list) -> pa.Table:
+    """An Arrow table shaped like a finelog result; types are inferred from the values."""
+    return pa.table(dict(columns))
+
+
 _ONE_ROW = finelog_result(t=[datetime(2026, 7, 17, 3, 0, tzinfo=UTC)], value=[1.0])
 
 
@@ -41,12 +47,8 @@ class FakeSource:
         return self._table
 
 
-def _client(source: FakeSource, **overrides) -> TestClient:
-    config = BridgeConfig(
-        max_rows=1000,
-        cache_ttl=overrides.get("cache_ttl", 20.0),
-        query_timeout_ms=5000,
-    )
+def _client(source: FakeSource, cache_ttl: float = 20.0) -> TestClient:
+    config = BridgeConfig(max_rows=1000, cache_ttl=cache_ttl, query_timeout_ms=5000)
     return TestClient(create_app(config, {"marin": source}))
 
 
@@ -54,20 +56,18 @@ def _get(client: TestClient, sql: str, **params):
     return client.get("/marin/query", params={"sql": sql, "from": FROM_MS, "to": TO_MS, **params})
 
 
-def test_query_runs_the_sql_and_returns_json_rows():
-    source = FakeSource(_ONE_ROW)
-    resp = _get(_client(source), 'SELECT t, value FROM "iris.task" WHERE ts >= {{from}} AND ts < {{to}}')
+def test_query_returns_json_rows_with_millis_timestamps():
+    resp = _get(_client(FakeSource(_ONE_ROW)), 'SELECT t, value FROM "iris.task" WHERE ts >= {{from}} AND ts < {{to}}')
     assert resp.status_code == 200
     assert resp.json() == [{"t": 1_784_257_200_000, "value": 1.0}]
 
 
-def test_query_substitutes_the_window_macros_before_running():
+def test_query_substitutes_window_macros_before_running():
     source = FakeSource(_ONE_ROW)
     _get(_client(source), "SELECT value FROM t WHERE ts >= {{from}} AND ts < {{to}}")
-    assert (
-        source.queries[0]
-        == "SELECT value FROM t WHERE ts >= TIMESTAMP '2026-07-17 03:00:00' AND ts < TIMESTAMP '2026-07-17 04:00:00'"
-    )
+    assert source.queries == [
+        "SELECT value FROM t WHERE ts >= TIMESTAMP '2026-07-17 03:00:00' AND ts < TIMESTAMP '2026-07-17 04:00:00'"
+    ]
 
 
 def test_missing_sql_is_a_400():
@@ -76,7 +76,7 @@ def test_missing_sql_is_a_400():
     assert "sql" in resp.json()["error"]
 
 
-def test_a_macro_without_its_bound_is_a_400():
+def test_macro_without_its_bound_is_a_400():
     resp = _client(FakeSource()).get("/marin/query", params={"sql": "SELECT 1 WHERE ts >= {{from}}"})
     assert resp.status_code == 400
     assert "no matching time bound" in resp.json()["error"]
@@ -87,30 +87,23 @@ def test_unknown_cluster_is_a_400_naming_the_valid_ones():
     assert resp.status_code == 400
     error = resp.json()["error"]
     assert "nope" in error and "marin" in error
-    assert not error.startswith('"')
 
 
-def test_oversized_result_is_a_400_telling_the_caller_what_to_do():
+def test_oversized_result_is_a_400_with_guidance():
     resp = _get(_client(FakeSource(raises=QueryResultTooLargeError("query returned 500000 rows"))), "SELECT 1")
     assert resp.status_code == 400
     assert "narrow the time range" in resp.json()["error"]
 
 
 def test_repeated_identical_panels_hit_finelog_once():
-    # A shared dashboard refreshing across N viewers must not multiply through to
-    # the finelog hub; Grafana's own query caching is Enterprise-only.
     source = FakeSource(_ONE_ROW)
     client = _client(source)
-    first = _get(client, "SELECT value FROM t WHERE ts >= {{from}} AND ts < {{to}}")
-    second = _get(client, "SELECT value FROM t WHERE ts >= {{from}} AND ts < {{to}}")
-    assert first.json() == second.json()
+    sql = "SELECT value FROM t WHERE ts >= {{from}} AND ts < {{to}}"
+    assert _get(client, sql).json() == _get(client, sql).json()
     assert len(source.queries) == 1
 
 
-def test_drifting_relative_window_still_hits_the_cache():
-    # Grafana sends a relative range as absolute millis, so both edges advance a
-    # little each refresh. The macros keep the SQL identical and the snapped bucket
-    # keeps the key stable, so it stays one query.
+def test_relative_window_drifting_within_the_ttl_stays_one_query():
     source = FakeSource(_ONE_ROW)
     client = _client(source, cache_ttl=60.0)
     sql = "SELECT value FROM t WHERE ts >= {{from}} AND ts < {{to}}"
@@ -123,11 +116,65 @@ def test_windows_further_apart_than_the_ttl_are_cached_separately():
     source = FakeSource(_ONE_ROW)
     client = _client(source, cache_ttl=20.0)
     sql = "SELECT value FROM t WHERE ts >= {{from}} AND ts < {{to}}"
-    _get(client, sql)
     hour = 3_600_000
+    _get(client, sql)
     _get(client, sql, **{"from": FROM_MS + hour, "to": TO_MS + hour})
     assert len(source.queries) == 2
 
 
+def test_json_labels_flatten_into_columns():
+    source = FakeSource(finelog_result(value=[3.0], labels=['{"region": "us-east5", "scope": "pool"}']))
+    assert _get(_client(source), "SELECT value, labels FROM t").json() == [
+        {"value": 3.0, "label_region": "us-east5", "label_scope": "pool"}
+    ]
+
+
+def test_unparseable_labels_cell_keeps_the_row():
+    # One malformed cell is schema drift; the panel still gets its row.
+    source = FakeSource(finelog_result(value=[1.0], labels=["{not json"]))
+    assert _get(_client(source), "SELECT value, labels FROM t").json() == [{"value": 1.0, "labels": "{not json"}]
+
+
 def test_health_lists_configured_clusters():
     assert _client(FakeSource()).get("/health").json() == {"status": "ok", "clusters": ["marin"]}
+
+
+def test_cache_coalesces_concurrent_misses_on_one_key():
+    # N callers racing a cold key compute once. Pin the order: the first caller is
+    # inside compute, holding the key lock, before the rest start.
+    cache: TtlCache[int] = TtlCache(ttl=60.0)
+    computing = threading.Event()
+    release = threading.Event()
+    calls: list[int] = []
+    results: list[int] = []
+
+    def compute():
+        calls.append(1)
+        computing.set()
+        release.wait(timeout=5)
+        return 7
+
+    def worker():
+        results.append(cache.get_or_compute("k", compute))
+
+    first = threading.Thread(target=worker)
+    first.start()
+    assert computing.wait(timeout=5), "first caller never entered compute"
+    others = [threading.Thread(target=worker) for _ in range(3)]
+    for t in others:
+        t.start()
+    release.set()
+    for t in [first, *others]:
+        t.join(timeout=10)
+
+    assert len(calls) == 1
+    assert results == [7, 7, 7, 7]
+
+
+def test_cache_prunes_expired_entries_on_write():
+    # Keys embed a rotating time bucket, so an insert-only cache grows without bound
+    # on a long-lived process. At ttl=0 every entry is stale on arrival.
+    cache: TtlCache[int] = TtlCache(ttl=0.0)
+    for i in range(50):
+        cache.get_or_compute(f"bucket-{i}", lambda i=i: i)
+    assert len(cache) == 0
